+++ controllers/__init__.py
@@ -10,6 +10,7 @@ from controllers.pins import PinsController, RulesController
 from controllers.questions import QuestionsController
 from controllers.reports import ReportsController
 from controllers.search import SearchController
+from controllers.suggestions import SuggestionsController
 from controllers.reminders import RemindersController
 from controllers.todos import TodosController
 from controllers.tools import ToolsController
@@ -19,4 +20,4 @@ CONTROLLERS = {c.resource: c for c in (RemindersController(), QuestionsControlle
                                           PinsController(), RulesController(), InboxController(),
                                           WorkController(), DocsController(),
                                           EnvironmentController(), SearchController(), ToolsController(),
-                                          ActivityController(), CommentsController(), ReportsController(), NotificationsController())}
+                                          ActivityController(), CommentsController(), ReportsController(), NotificationsController(), SuggestionsController())}